Refactor utilities and fix build
